pyproject.toml
[tool.poetry]
name = "pyhazel"
version = "0.1.0"
description = ""
authors = ["twje"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.23.3"
glfw = "^2.5.3"
PyOpenGL = "^3.1.6"
imgui = "^1.4.1"
PyGLM = "^2.5.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

